+++ src/Form.js
@@ -148,13 +148,13 @@ export default class Form extends Element {
     else {
       this.element = null;
     }
+    this.options = formOptions;
+    this.options.events = this.events;
     if (form) {
       this.setForm(form)
         .then(() => this.readyResolve(this.instance))
         .catch(this.readyReject);
     }
-    this.options = formOptions;
-    this.options.events = this.events;
     this.display = '';
   }
 
@@ -321,7 +321,7 @@ export default class Form extends Element {
               } else {
                 this.instance.url = formParam;
               }
-              
+
               this.instance.nosubmit = false;
               this._form = this.instance.form = form;
               if (submission) {