+++ src/components/datagrid/DataGrid.js
@@ -516,7 +516,7 @@ export default class DataGridComponent extends NestedArrayComponent {
 
   removeRow(index) {
     const makeEmpty = index === 0 && this.rows.length === 1;
-    const flags = { isReordered: !makeEmpty, resetValue: makeEmpty };
+    const flags = { isReordered: !makeEmpty, resetValue: makeEmpty, modified: true };
     this.splice(index, flags);
     this.emit('dataGridDeleteRow', { index });
     const [row] = this.rows.splice(index, 1);